feat: move k8s client wrappers to k8s.clients
