+++ components/renku_data_services/k8s_watcher/db.py
@@ -2,107 +2,19 @@
 
 from __future__ import annotations
 
-import contextlib
 import logging
 from collections.abc import AsyncIterable, Awaitable, Callable
-from typing import Any
 
-import kr8s
 import sqlalchemy
 from sqlalchemy import bindparam, select
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from renku_data_services.errors import errors
-from renku_data_services.k8s.models import APIObjectInCluster, Cluster, ClusterId, K8sObject, K8sObjectMeta, ListFilter
+from renku_data_services.k8s.models import APIObjectInCluster, K8sObject, K8sObjectMeta, ListFilter
 from renku_data_services.k8s_watcher.orm import K8sObjectORM
 
 type EventHandler = Callable[[APIObjectInCluster], Awaitable[None]]
-
-
-class K8sClient:
-    """A wrapper around a kr8s k8s client, acts on all resources over many clusters."""
-
-    def __init__(self, clusters: dict[ClusterId, Cluster]) -> None:
-        self.__clusters = clusters
-
-    def __get_cluster_or_die(self, cluster_id: ClusterId) -> Cluster:
-        cluster = self.__clusters.get(cluster_id)
-        if not cluster:
-            raise errors.MissingResourceError(
-                message=f"Could not find cluster with id {cluster_id} in the list of clusters."
-            )
-        return cluster
-
-    async def create(self, obj: K8sObject) -> K8sObject:
-        """Create the k8s object."""
-        cluster = self.__get_cluster_or_die(obj.cluster)
-        api_obj = APIObjectInCluster.from_k8s_object(obj, cluster.api)
-        await api_obj.obj.create()
-        return api_obj.meta.with_manifest(api_obj.obj.to_dict())
-
-    async def patch(self, meta: K8sObjectMeta, patch: dict[str, Any] | list[dict[str, Any]]) -> K8sObject:
-        """Patch a k8s object.
-
-        If the patch is a list we assume that we have a rfc6902 json patch like
-        `[{ "op": "add", "path": "/a/b/c", "value": [ "foo", "bar" ] }]`.
-        If the patch is a dictionary then it is considered to be a rfc7386 json merge patch.
-        """
-        obj = await self._get(meta)
-        if not obj:
-            raise errors.MissingResourceError(message=f"The k8s resource with metadata {meta} cannot be found.")
-        patch_type = "json" if isinstance(patch, list) else None
-        await obj.obj.patch(patch, type=patch_type)
-        return meta.with_manifest(obj.obj.to_dict())
-
-    async def delete(self, meta: K8sObjectMeta) -> None:
-        """Delete a k8s object."""
-        obj = await self._get(meta)
-        if not obj:
-            return None
-        with contextlib.suppress(kr8s.NotFoundError):
-            await obj.obj.delete(propagation_policy="Foreground")
-
-    async def _get(self, meta: K8sObjectMeta) -> APIObjectInCluster | None:
-        return await anext(aiter(self.__list(meta.to_list_filter())), None)
-
-    async def get(self, meta: K8sObjectMeta) -> K8sObject | None:
-        """Get a specific k8s object, None is returned if the object does not exist."""
-        obj = await self._get(meta)
-        if not obj:
-            return None
-        return meta.with_manifest(obj.obj.to_dict())
-
-    async def __list(self, filter: ListFilter) -> AsyncIterable[APIObjectInCluster]:
-        clusters = list(self.__clusters.values())
-        if filter.cluster:
-            single_cluster = self.__clusters.get(filter.cluster)
-            clusters = [single_cluster] if single_cluster else []
-        for cluster in clusters:
-            if filter.namespace is not None and filter.namespace != cluster.namespace:
-                continue
-            names = [filter.name] if filter.name else []
-
-            try:
-                res = await cluster.api.async_get(
-                    filter.kind,
-                    *names,
-                    label_selector=filter.label_selector,
-                    namespace=filter.namespace,
-                )
-            except (kr8s.ServerError, kr8s.APITimeoutError):
-                continue
-
-            if not isinstance(res, list):
-                res = [res]
-            for r in res:
-                yield APIObjectInCluster(r, cluster.id)
-
-    async def list(self, filter: ListFilter) -> AsyncIterable[K8sObject]:
-        """List all k8s objects."""
-        results = self.__list(filter)
-        async for r in results:
-            yield r.to_k8s_object()
 
 
 class K8sDbCache:
@@ -195,67 +107,3 @@ class K8sDbCache:
                 )
             async for res in await session.stream_scalars(stmt):
                 yield res.dump()
-
-
-class CachedK8sClient(K8sClient):
-    """A wrapper around a kr8s k8s client.
-
-    Provides access to a cache for listing and reading resources but fallback to the cluster for other operations.
-    """
-
-    def __init__(self, clusters: dict[ClusterId, Cluster], cache: K8sDbCache, kinds_to_cache: list[str]) -> None:
-        super().__init__(clusters)
-        self.cache = cache
-        self.__kinds_to_cache = [k.lower() for k in kinds_to_cache]
-
-    async def create(self, obj: K8sObject) -> K8sObject:
-        """Create the k8s object."""
-        if obj.meta.kind.lower() in self.__kinds_to_cache:
-            await self.cache.upsert(obj)
-        try:
-            obj = await super().create(obj)
-        except:
-            # if there was an error creating the k8s object, we delete it from the db again to now have ghost entries
-            if obj.meta.kind.lower() in self.__kinds_to_cache:
-                await self.cache.delete(obj)
-            raise
-        return obj
-
-    async def patch(self, meta: K8sObjectMeta, patch: dict[str, Any] | list[dict[str, Any]]) -> K8sObject:
-        """Patch a k8s object."""
-        obj = await super().patch(meta, patch)
-        if meta.kind.lower() in self.__kinds_to_cache:
-            await self.cache.upsert(obj)
-        return obj
-
-    async def delete(self, meta: K8sObjectMeta) -> None:
-        """Delete a k8s object."""
-        await super().delete(meta)
-        if meta.kind.lower() in self.__kinds_to_cache:
-            await self.cache.delete(meta)
-
-    async def get(self, meta: K8sObjectMeta) -> K8sObject | None:
-        """Get a specific k8s object, None is returned if the object does not exist."""
-        if meta.kind.lower() in self.__kinds_to_cache:
-            res = await self.cache.get(meta)
-        else:
-            res = await super().get(meta)
-        if res is None:
-            return None
-        return res
-
-    async def get_api_object(self, meta: K8sObjectMeta) -> APIObjectInCluster | None:
-        """Get a kr8s object directly, bypassing the cache.
-
-        Note: only use this if you actually need to do k8s operations.
-        """
-        res = await super()._get(meta)
-        if res is None:
-            return None
-        return res
-
-    async def list(self, filter: ListFilter) -> AsyncIterable[K8sObject]:
-        """List all k8s objects."""
-        results = self.cache.list(filter) if filter.kind.lower() in self.__kinds_to_cache else super().list(filter)
-        async for res in results:
-            yield res
